fix(ci): keep the unit suite hermetic — extract pure helpers
CI went red on PR #3 and it was mine. The tests I added in 836cf7f and
914bce3 import ./client and ./propose, whose import chains reach
src/config.ts — and that module ends in `export const config = loadConfig()`,
which THROWS when DATABASE_URL is absent. Vitest loads .env locally, so all
21 passed on my machine; the runner has no .env, so importing the module
under test was itself the failure.

ci.yml states the contract: "hermetic, Google APIs are mocked; no DB needed."
Setting a dummy DATABASE_URL in the workflow would have gone against that,
so the config dependency gets dropped instead of accommodated.

Both helpers were pure functions sitting in modules that talk to prisma and
OAuth, and both are better off in their own:

  - src/drive/retry.ts — isTransientTransportError + withTransientRetry.
    Only depends on the logger (pino, no config). The wrap-don't-nest
    rationale travels with the code.
  - src/scan/note-marker.ts — the candidate marker format. It's a contract
    BETWEEN propose.ts (writer) and process-batch.ts (reader), so a module
    both import is a truer home than one importing from the other.

No behavior change. Tests renamed to match and verified the way CI runs
them: .env moved aside, 21/21 green.
